Configure Docker
Add Dockerfile and docker-compose.yml to configure a 'jekyll' service based on the starefossen/github-pages Docker image with project dependencies installed, with port 4000 on the host linked to port 4000 on the container, and with the current working directory on the host mapped to the /usr/src/app volume on the container.

Update _config.yml to exclude non-build files from the Jekyll site.

Add .gitignore to exclude Ruby build artifacts from the repository.